requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==3.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.5.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.20
